Improve nucleus controller
