+++ app/src/main/java/eu/kanade/tachiyomi/ui/base/controller/NucleusController.kt
@@ -1,0 +1,19 @@
+package eu.kanade.tachiyomi.ui.base.controller
+
+import eu.kanade.tachiyomi.ui.base.presenter.NucleusConductorDelegate
+import eu.kanade.tachiyomi.ui.base.presenter.NucleusConductorLifecycleListener
+import nucleus.factory.PresenterFactory
+import nucleus.presenter.Presenter
+
+@Suppress("LeakingThis")
+abstract class NucleusController<P : Presenter<*>> : RxController(), PresenterFactory<P> {
+
+    private val delegate = NucleusConductorDelegate(this)
+
+    val presenter: P
+        get() = delegate.presenter
+
+    init {
+        addLifecycleListener(NucleusConductorLifecycleListener(delegate))
+    }
+}
